Teamspace/requirements.txt
opencv-python
torch
torchvision
Pillow
requests
numpy
kneron-sdk